Working break/gravity :D

+++ src/main/scala/exp/draw/Draw2D.scala
@@ -5,10 +5,10 @@ import common.immutable.{Cube, Pos}
 import scala.swing.Color
 
 class Draw2D(g: scala.swing.Graphics2D, val view: View2D) {
-  def fillRect(x: Cube[Double], color: Color): Unit = window.fillRect(view.scale(x), color)
-  def lineRect(x: Cube[Double], color: Color): Unit = window.lineRect(view.scale(x), color)
+  def fillRect(x: Cube[Long], color: Color): Unit = window.fillRect(view.scale(x), color)
+  def lineRect(x: Cube[Long], color: Color): Unit = window.lineRect(view.scale(x), color)
 
-  def drawText(str: String, pos: Pos[Double], color: Color = Draw2D.kBlack): Unit = window.drawText(str, view.scale(pos), color)
+  def drawText(str: String, pos: Pos[Long], color: Color = Draw2D.kBlack): Unit = window.drawText(str, view.scale(pos), color)
 
   // Drawing directly in window coordinates
   object window {
