+++ src/main/scala/exp/material/Liquid.scala
@@ -1,7 +1,7 @@
 package exp.material
 
-abstract class Liquid(_name: String, _tension: Double) extends Material(_name) {
+abstract class Liquid(_name: String, _tension: Long) extends Material(_name) {
   override def liquid: Boolean = true
-  val tension: Double = _tension
-  lazy val tension2: Double = tension*tension
+  val tension: Long = _tension
+  lazy val tension2: Long = tension*tension
 }
